`Refactor saveFileAndScreenshot to conversion-util`
Moved the `saveFileAndScreenshot` function from `ExploreModeAnthropicProvider` to `conversion-util` to improve modularity and reusability. Updated references in `ExploreModeAnthropicProvider` to use the newly centralized function.
